pt3/requirements.txt
emoji==2.10.1
mysql-connector-python
flask